Necessary changes to build with autotools
This change introduces the autoconf and automake inputs with the necessary
changes to get a working build with the standard autotools dance.

Some of the cli targets might be incomplete (test_cli.exe?). Not sure these are
needed yet (if ever). The original Makefile has been updated and kept in place,
so building with the static Makefile should still work for now.