Generic Fwk::LinkedList implementation.
Moved RoutingTable's internal linked list implementation to the generic
Fwk::LinkedList.

Fwk::LinkedList is an invasive collection. It takes in a single template
parameter, NodeType, that represents the type of node that can be inserted into
the collection. NodeType must be the name of a class that inherits from
Fwk::LinkedList::Node.