Calculate buffer offset in Packet
Because the storage buffer for a Packet is a PacketBuffer, and PacketBuffers
may resize, we cannot keep a buffer offset relative to the start of the buffer,
as it may change. Instead, we keep a reverse offset, an offset relative to the
end of the buffer, and calculate the forward offset from that. PacketBuffers
only extend at the head, so a reverse offset will never change.