Avoid using enclosingPacket()
Using enclosingPacket() here makes the incorrect assumption that the enclosing
packet will always be an EthernetPacket. This may not be the case when we just
created a GRE packet. Instead, what we want to do is *add* an Ethernet header,
so the code now does just that in the appropriate way.