BaseNotifier/ee inherit from NamedInterface
This is necessary to avoid the diamond problem if a class wishes to be a
notifier and inherit from NamedInterface. In such a case, both would inherit
from two templatized versions of PtrInterface, a problem which I do not believe
is resolvable with virtual inheritance.

BaseNotifier/BaseNotifiee now inherit from NamedInterface, but still provide a
no-argument constructor, so a name becomes optional.